base_agent.py
# Base Agent - Remittance & Exchange Advisor
import os
import sys
from google import genai
from dotenv import load_dotenv

load_dotenv()

client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

query = sys.argv[1]

response = client.models.generate_content(
        model="gemini-2.5-flash",
        contents=query
)

print (response.text)
